pyproject.toml
[tool.poetry]
name = "tuna"
version = "0.1.0"
description = ""
authors = ["dnl-strnl <129630837+dnl-strnl@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
hydra-core = "^1.3.2"
mlx = "^0.5.0"
numpy = "^1.26.4"
transformers = "^4.38.2"
huggingface-hub = "^0.21.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
